requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
atproto==0.0.59
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
exceptiongroup==1.2.2
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.51
langchain-ollama==0.3.0
langchain-text-splitters==0.3.8
langsmith==0.3.24
libipld==3.0.1
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvtx==0.2.11
oauthlib==3.2.2
ollama==0.4.7
orjson==3.10.16
packaging==24.2
pillow==11.1.0
praw==7.8.1
prawcore==2.4.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sniffio==1.3.1
snscrape==0.7.0.20230622
soupsieve==2.6
SQLAlchemy==2.0.40
sympy==1.13.1
tenacity==9.1.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
triton==3.2.0
tweepy==4.15.0
typing_extensions==4.13.0
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
websockets==13.1
zstandard==0.23.0
